Reorganized some test utils for general use
